+++ crates/ra_flycheck/Cargo.toml
@@ -9,12 +9,8 @@ doctest = false
 
 [dependencies]
 crossbeam-channel = "0.4.0"
-lsp-types = { version = "0.74.0", features = ["proposed"] }
 log = "0.4.8"
 cargo_metadata = "0.10.0"
 serde_json = "1.0.48"
 jod-thread = "0.1.1"
 ra_toolchain = { path = "../ra_toolchain" }
-
-[dev-dependencies]
-insta = "0.16.0"
