+++ crates/rust-analyzer/Cargo.toml
@@ -29,6 +29,7 @@ rustc-hash = "1.1.0"
 serde = { version = "1.0.106", features = ["derive"] }
 serde_json = "1.0.48"
 threadpool = "1.7.1"
+cargo_metadata = "0.10.0"
 
 stdx = { path = "../stdx" }
 
@@ -53,6 +54,7 @@ winapi = "0.3.8"
 
 [dev-dependencies]
 tempfile = "3.1.0"
+insta = "0.16.0"
 test_utils = { path = "../test_utils" }
 
 [features]
